Extract generic RevertToRevisionAction
