+++ wagtail/actions/revert_to_page_revision.py
@@ -1,6 +1,7 @@
-from django.core.exceptions import PermissionDenied
-
-from wagtail.actions.base import BaseAction
+from wagtail.actions.revert_to_revision import (
+    RevertToRevisionAction,
+    RevertToRevisionPermissionError,
+)
 
 
 class RevertToPageRevisionError(RuntimeError):
@@ -11,7 +12,7 @@ class RevertToPageRevisionError(RuntimeError):
     pass
 
 
-class RevertToPageRevisionPermissionError(PermissionDenied):
+class RevertToPageRevisionPermissionError(RevertToRevisionPermissionError):
     """
     Raised when the revision revert cannot be performed due to insufficient permissions.
     """
@@ -19,30 +20,9 @@ class RevertToPageRevisionPermissionError(PermissionDenied):
     pass
 
 
-class RevertToPageRevisionAction(BaseAction):
-    action_name = "revert"
-    permission_error_class = RevertToPageRevisionPermissionError
-
-    def __init__(
-        self,
-        page,
-        revision,
-        user=None,
-        log_action="wagtail.revert",
-        approved_go_live_at=None,
-        changed=True,
-        clean=True,
-    ):
-        super().__init__(page, user=user)
-        self.page = page
-        self.revision = revision
-        self.log_action = log_action
-        self.approved_go_live_at = approved_go_live_at
-        self.changed = changed
-        self.clean = clean
-
+class RevertToPageRevisionAction(RevertToRevisionAction):
     def check(self, skip_permission_checks=False):
-        if self.page.alias_of_id:
+        if self.instance.alias_of_id:
             raise RevertToPageRevisionError(
                 "Revisions are not required for alias pages as they are an exact copy of another page."
             )
@@ -51,20 +31,8 @@ class RevertToPageRevisionAction(BaseAction):
         if (
             self.user
             and not skip_permission_checks
-            and not self.page.permissions_for_user(self.user).can_edit()
+            and not self.instance.permissions_for_user(self.user).can_edit()
         ):
             raise RevertToPageRevisionPermissionError(
                 "You do not have permission to edit this page."
             )
-
-    def execute(self, skip_permission_checks=False):
-        self.check(skip_permission_checks=skip_permission_checks)
-
-        return self.revision.as_object().save_revision(
-            previous_revision=self.revision,
-            user=self.user,
-            log_action=self.log_action,
-            approved_go_live_at=self.approved_go_live_at,
-            changed=self.changed,
-            clean=self.clean,
-        )
